Rename crawler files
